also move cocktail-dkg ciphersuite tests to separate file; to be compatible with gencode
