+++ frost-ed448/tests/integration_tests.rs
@@ -14,13 +14,6 @@ fn check_sign_with_dkg() {
     frost_core::tests::ciphersuite_generic::check_sign_with_dkg::<Ed448Shake256, _>(rng);
 }
 
-#[cfg(feature = "cocktail-dkg")]
-#[test]
-fn check_sign_with_cocktail_dkg() {
-    let rng = rand::rngs::OsRng;
-
-    frost_core::tests::cocktail_dkg::check_sign_with_cocktail_dkg::<Ed448Shake256, _>(rng);
-}
 
 #[test]
 fn check_dkg_part1_fails_with_invalid_signers_min_signers() {
@@ -315,26 +308,3 @@ async fn check_async_sign_with_dealer() {
     .unwrap();
 }
 
-#[cfg(feature = "cocktail-dkg")]
-#[test]
-fn check_cocktail_dkg_test_vectors() {
-    use sha3::{
-        digest::{ExtendableOutput, Update, XofReader},
-        Shake256,
-    };
-
-    let json_str = include_str!("helpers/cocktail-dkg-ed448-shake256.json");
-
-    frost_core::tests::cocktail_dkg::check_cocktail_dkg_test_vectors::<Ed448Shake256, _>(
-        json_str,
-        |data| {
-            let mut h = Shake256::default();
-            h.update(data);
-            let mut out = vec![0u8; 64];
-            h.finalize_xof().read(&mut out);
-            out
-        },
-        false, // encrypted shares: 73 vs 72 byte format mismatch
-        false, // recovery: ciphertext format incompatible
-    );
-}
