+++ frost-p256/tests/integration_tests.rs
@@ -14,13 +14,6 @@ fn check_sign_with_dkg() {
     frost_core::tests::ciphersuite_generic::check_sign_with_dkg::<P256Sha256, _>(rng);
 }
 
-#[cfg(feature = "cocktail-dkg")]
-#[test]
-fn check_sign_with_cocktail_dkg() {
-    let rng = rand::rngs::OsRng;
-
-    frost_core::tests::cocktail_dkg::check_sign_with_cocktail_dkg::<P256Sha256, _>(rng);
-}
 
 #[test]
 fn check_dkg_part1_fails_with_invalid_signers_min_signers() {
@@ -312,17 +305,3 @@ async fn check_async_sign_with_dealer() {
     .unwrap();
 }
 
-#[cfg(feature = "cocktail-dkg")]
-#[test]
-fn check_cocktail_dkg_test_vectors() {
-    use sha2::{Digest, Sha256};
-
-    let json_str = include_str!("helpers/cocktail-dkg-p256-sha256.json");
-
-    frost_core::tests::cocktail_dkg::check_cocktail_dkg_test_vectors::<P256Sha256, _>(
-        json_str,
-        |data| Sha256::digest(data).to_vec(),
-        true, // encrypted shares match (XAES-256-GCM)
-        true, // recovery is tested
-    );
-}
